Sensor data with LAN webpage.ino
In this code, we read rain status from a rain sensor module and temperature and humidity data from a DHT11 sensor. Then we send the data to the cloud through HTTPS push protocol on a Local Area Network.  All computers connected on the LAN can access the updated data using the web address